# ♻️  REFACTOR: move basic parsing into `BaseParser` class