--- a/src/aiida_quantumespresso/calculations/namelists.py
+++ b/src/aiida_quantumespresso/calculations/namelists.py
@@ -61,9 +61,9 @@ class NamelistsCalculation(CalcJob):
         spec.exit_code(302, 'ERROR_OUTPUT_STDOUT_MISSING',
             message='The retrieved folder did not contain the required stdout output file.')
         spec.exit_code(310, 'ERROR_OUTPUT_STDOUT_READ',
-            message='The stdout output file could not be read.')
+            message='An exception was raised while reading the `stdout` file: {exception}')
         spec.exit_code(311, 'ERROR_OUTPUT_STDOUT_PARSE',
-            message='The stdout output file could not be parsed.')
+            message='An exception was raised while parsing the `stdout` file: {exception}')
         spec.exit_code(312, 'ERROR_OUTPUT_STDOUT_INCOMPLETE',
             message='The stdout output file was incomplete probably because the calculation got interrupted.')
         # yapf: enable