--- a/tests/parsers/test_ph/test_ph_default_default_.yml
+++ b/tests/parsers/test_ph/test_ph_default_default_.yml
@@ -1,3 +1,4 @@
+code_version: v.6.1
 dielectric_constant:
 - - 57.36256076907993
   - -2.842170943040401e-14