… src/aiida_quantumespresso/parsers/open_grid.py
… src/aiida_quantumespresso/parsers/parse_raw/base.py
… src/aiida_quantumespresso/parsers/parse_raw/cp.py
… src/aiida_quantumespresso/parsers/parse_raw/neb.py
… src/aiida_quantumespresso/parsers/parse_raw/ph.py
… src/aiida_quantumespresso/parsers/parse_raw/pw2gw.py
… src/aiida_quantumespresso/parsers/ph.py
… src/aiida_quantumespresso/parsers/pp.py
… src/aiida_quantumespresso/parsers/projwfc.py
… src/aiida_quantumespresso/parsers/pw.py
… src/aiida_quantumespresso/parsers/pw2gw.py
… src/aiida_quantumespresso/parsers/pw2wannier90.py
… src/aiida_quantumespresso/parsers/q2r.py
… src/aiida_quantumespresso/parsers/xspectra.py
… tests/parsers/test_cp/test_cp_default_6_6_autopilot_.yml
… tests/parsers/test_cp/test_cp_default_6_6_cgstep_.yml
… tests/parsers/test_cp/test_cp_default_6_6_cgsteps_.yml
… tests/parsers/test_cp/test_cp_default_6_6_verlet_.yml
… tests/parsers/test_cp/test_cp_default_default_.yml
… tests/parsers/test_dos/test_dos_default.yml
… tests/parsers/test_matdyn/test_matdyn_default.yml
… tests/parsers/test_neb.py
… tests/parsers/test_neb/test_neb_default.yml
… tests/parsers/test_open_grid/test_open_grid_default.yml
… tests/parsers/test_ph.py
… tests/parsers/test_ph/test_ph_default_no_modes_printed_.yml
… tests/parsers/test_ph/test_ph_default_single_qpoint_.yml
… tests/parsers/test_ph/test_ph_not_converged.yml
… tests/parsers/test_ph/test_ph_out_of_walltime.yml
… tests/parsers/test_pw2gw/test_pw2gw_default_data.yml