Fixes #750 
Fixes #797

Large refactoring of the basic parsing features to deduplicate code and rely on regular expressions via `re` instead of iterations over lines. The main changes are:

* The `Parser` class is renamed to `BaseParser` class to avoid confusion with the `aiida.core` `Parser` class.
* The `BaseParser.parse_stdout_from_retrieved` method can be used by all subclasses to obtain the `stdout`, basic parsed data and logs from retrieving and parsing the `stdout`.
* The error and warning messages are defined as class attributes. Base error messages are defined on the `BaseParser` class, which can be extended on the subclasses via the `class_[error,warning]_map` attributes.
* A `BaseParser.check_base_errors()` method is added, which checks for basic errors that indicate the `stdout` cannot be parsed for more data.
* Shared exit codes are moved to the `NamelistsCalculation` spec from that of the subclasses.
* Several changes in the logic of returning exit codes and emitting the logs. Some of the discussions on these topics have not been fully concluded, and will be revisited when integrating the `BaseParser` into the `PwParser`, see:
  https://github.com/aiidateam/aiida-quantumespresso/issues/953