@mbercx 2022-01-27
Some timing information for a very long `aiida.out` file (~140k lines) without errors:

* Old `parse_output_base`:

```
14.4 ms ± 313 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

* New `parse_output_base`:

```
5.14 ms ± 68.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

and a very short (131 lines) file with errors:

* Old `parse_output_base`:

```
11.3 µs ± 231 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

* New `parse_output_base`:

```
48.1 µs ± 815 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```