@mbercx 2023-05-07
@sphuber ok, I refactored my initial refactor quite a bit, so perhaps this needs a full review again, apologies. It was also quite a bit of work to fix all the merge conflicts after rebase, so I would now really push to get this PR in.

A couple of quick notes that I decided upon:

* The code version in the `output_parameters` shouldn't have a `v.` prefix. It's redundant.
* The `wall_time` output is similarly redundant and less useful than `wall_time_seconds`.

Now a more elaborate design discussion on the changes here and TODO's for parsers.

## Keeping the `parse` method readable

Understanding what the parser parses when and which exit codes it returns at what point is difficult with many of the parser classes. We want to make sure that:

1. The "best" exit code is returned, i.e. the one that is most specific or allows for useful error handling.
2. As much data as possible is parsed and added to the outputs of the calculation job.

To improve clarity and make these two requirements possible, I think the `parse` method should:

1. Not contain any _actual_ parsing. Any real parsing should happen in a separate method.
3. Instantiate a single logging container that is passed _through_ these parsing methods.
4. Take care of decisions on when to return which exit code.
5. Output data at the appropriate time.

Note that this PR _does not_ fully implement these ideas for all parser classes. It's already big enough as it is. But it already shows how this would be possible for e.g. the `open_grid.x` parser.

## Other improvements 

* **Exit code registry**: Although this PR already make an effort of making sure exit codes are consistent across the calculation job classes, it's still tricky to keep track of all the exit codes they are used any why they are there or have a certain number. I think we should create an (automated) exit code registry in the documentation where we keep track of all this. Although two calculation jobs can in principle have the same exit code for different issues, we should probably also avoid this.

* **Stop using strings for exit codes**: Now we use strings to keep track of the exit codes. I think this is a bit fragile and doesn't allow for nice features like tab-completion. It would be better to have a different approach, but I'd have to think a bit on how to do this.

* `PhCalculation` and `PpCalculation` aren't subclasses from `NamelistCalculation`, and hence have to redefine the basic exit codes. So do `PwCalculation` and `CpCalculation`. There may be reasons to treat these calculation jobs differently, but we should investigate if we can't further consolidate this calculation job hierarchy.