@mbercx 2023-05-10
Because of the size of the PR, and the fact that even though the tests pasts it can definitely break things, perhaps:

* @PNOGillespie @superstar54 maybe you can have a look a the XSpectra parsing?
* @qiaojunfeng can you have a look at the `open_grid.x` parsing changes?