@qiaojunfeng 2023-05-11
>     * @qiaojunfeng can you have a look at the `open_grid.x` parsing changes?

Had a quick look and seems good to me!
Anyhow `open_grid` is very simple so if the tests passed then I wouldn't worry.