@PNOGillespie 2023-05-12
Hi @mbercx, I'm running some live-fire tests of XSpectra to check for issues. Tests with the `XspectraCrystalWorkChain` appear to be fine and the `output_parameters` are as expected, I'm now setting up and re-running tests of the XSpectra-specific error codes to check those too.