@PNOGillespie 2023-05-12
@mbercx. To update the previous comment: I've tested all the XSpectra-specific exit codes, and they all work fine as well. The WorkChains themselves also rely on information in the `output_parameters` so, since the `WorkChains` run without issues, I will say that we're good for XSpectra-related stuff.