@mbercx 2023-06-08
> @mbercx did you want to get this merged before the release? If so, is there anything to be done from your side? Otherwise I will approve. We can deal with the open points later. I also would like to have this done so I can finalize my refactor PR for `PwParser`.

I've opened https://github.com/aiidateam/aiida-quantumespresso/issues/953 to keep track of the discussions that are left here. I think it makes sense to simply reconsider these decisions when integrating the `BaseParser` in the `PwParser`, we'll have a clearer idea of what to do during this process, I think.