@mbercx 2023-06-22
> want to finish that one up and release it?

Yessir! Sorry for the slow response.

> There were API changes, as far as I can see.

The internal parsing API was changed quite a bit, but I think the vast majority of the changes are backwards incompatible. Looking at the data regression files for the tests, the only changes in parsing output is the `code_version` and `wall_time_seconds`. Which API changes are you worried about breaking e.g. the QEapp?